+++ modules/model/src/main/java/no/ntnu/remotecode/model/DTO/kafkamessage/TaskAck.java
@@ -5,6 +5,7 @@ import no.ntnu.remotecode.model.enums.TaskStatus;
 
 @Data
 public class TaskAck {
-    int id;
+
+    long id;
     TaskStatus taskStatus;
 }
